# docs/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "APILFBI",
    "description": "API de gestión documental CRM ↔ Laserfiche 11.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5080/"
    }
  ],
  "paths": {
...
    "/api/v1/tipos-expediente/{idTipoExpediente}/tipos-documento": {
      "get": {
        "tags": [
          "TiposExpediente"
        ],
        "summary": "Criterio 3: tipos de documento de un tipo de expediente. tipoCliente (N/J) es opcional.",
        "parameters": [
          {
            "name": "idTipoExpediente",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "tipoCliente",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfTiposDocumentoDeExpedienteDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfTiposDocumentoDeExpedienteDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfTiposDocumentoDeExpedienteDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tipos-expediente/tipos-documento": {
      "get": {
        "tags": [
          "TiposExpediente"
        ],
        "summary": "Criterio 3 por nombre (o código) del tipo de expediente.",
        "parameters": [
          {
            "name": "nombreTipoExpediente",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoCliente",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfTiposDocumentoDeExpedienteDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfTiposDocumentoDeExpedienteDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfTiposDocumentoDeExpedienteDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tipos-expediente/{idTipoExpediente}/llaves": {
      "get": {
        "tags": [
          "TiposExpediente"
        ],
        "summary": "Llaves que el CRM debe enviar en POST /expedientes para este tipo de expediente.",
        "parameters": [
          {
            "name": "idTipoExpediente",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "tipoCliente",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfIReadOnlyListOfLlavesPorTipoClienteDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfIReadOnlyListOfLlavesPorTipoClienteDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaApiOfIReadOnlyListOfLlavesPorTipoClienteDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DocumentoDeTipoExpedienteDto": {
        "required": [
          "idTipoDocumento",
          "nombreTipoDocumento",
          "obligatorio",
          "aplicaA"
        ],
        "type": "object",
        "properties": {
          "idTipoDocumento": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombreTipoDocumento": {
            "type": "string"
          },
          "obligatorio": {
            "type": "boolean"
          },
          "aplicaA": {
            "type": "string"
          }
        }
      },
...
      "LlaveConfiguradaDto": {
        "required": [
          "codigo",
          "nombre",
          "obligatoria",
          "grupoIdentificacion",
          "prioridadGrupo",
          "tipoDato",
          "longitudMaxima",
          "expresionValidacion",
          "valoresPermitidos"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "obligatoria": {
            "type": "boolean"
          },
          "grupoIdentificacion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "prioridadGrupo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tipoDato": {
            "type": "string"
          },
          "longitudMaxima": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "expresionValidacion": {
            "type": [
              "null",
              "string"
            ]
          },
          "valoresPermitidos": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "LlavesPorTipoClienteDto": {
        "required": [
          "tipoCliente",
          "llaves"
        ],
        "type": "object",
        "properties": {
          "tipoCliente": {
            "type": "string"
          },
          "llaves": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LlaveConfiguradaDto"
            }
          }
        }
      },
...
      "RespuestaApiOfIReadOnlyListOfLlavesPorTipoClienteDto": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/LlavesPorTipoClienteDto"
            }
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
...
      "RespuestaApiOfTiposDocumentoDeExpedienteDto": {
        "required": [
          "codigo",
          "mensaje",
          "correlationId",
          "data"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mensaje": {
            "type": "string"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TiposDocumentoDeExpedienteDto"
              }
            ]
          }
        },
        "description": "Sobre estándar de las respuestas exitosas."
      },
...
      "TiposDocumentoDeExpedienteDto": {
        "required": [
          "idTipoExpediente",
          "nombreTipoExpediente",
          "tiposDocumento"
        ],
        "type": "object",
        "properties": {
          "idTipoExpediente": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombreTipoExpediente": {
            "type": "string"
          },
          "tiposDocumento": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentoDeTipoExpedienteDto"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "OAuth 2.0 client_credentials. El token dura 1 hora.",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "/api/v1/auth/token",
            "scopes": {
              "catalogos.leer": "catalogos.leer",
              "expedientes.escribir": "expedientes.escribir",
              "documentos.leer": "documentos.leer",
              "documentos.estado": "documentos.estado",
              "bitacora.leer": "bitacora.leer"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "catalogos.leer",
        "expedientes.escribir",
        "documentos.leer",
        "documentos.estado",
        "bitacora.leer"
      ]
    }
  ],
...
}